refactor: introduce quest domain/application boundaries
