+++ daedalus-dialog-editor/src/renderer/quest/application/index.ts
@@ -1,0 +1,2 @@
+export { QuestEditingService } from './QuestEditingService';
+export type { QuestCommandContext } from './QuestEditingService';
