+++ daedalus-dialog-editor/src/renderer/quest/domain/analysis.ts
@@ -1,0 +1,6 @@
+export {
+  analyzeQuest,
+  findDialogNameForFunction,
+  getQuestReferences,
+  getUsedQuestTopics
+} from '../../components/QuestEditor/questAnalysis';
