+++ daedalus-dialog-editor/src/renderer/quest/domain/commands.ts
@@ -1,0 +1,2 @@
+export { executeQuestGraphCommand } from '../../components/QuestEditor/commands';
+export type { QuestGraphCommand } from '../../components/QuestEditor/commands';
