+++ daedalus-dialog-editor/src/renderer/quest/domain/graph.ts
@@ -1,0 +1,1 @@
+export { buildQuestGraph } from '../../components/QuestEditor/questGraphUtils';
