+++ daedalus-dialog-editor/src/renderer/quest/domain/guardrails.ts
@@ -1,0 +1,6 @@
+export {
+  analyzeQuestGuardrails,
+  getNewQuestGuardrailWarnings,
+  getQuestGuardrailDeltaWarnings,
+  isQuestGuardrailWarningBlocking
+} from '../../components/QuestEditor/questGuardrails';
